fix(dataset): 优化数据集更新和检索功能
- 优化 DocumentService 中的更新操作，确保返回更新后的最新实体
- 改进 FullTextSearchService 中的全文搜索算法，提高搜索结果准确性- 调整 ParagraphVO 中的相似度分数类型，从 Double 改为 Float
- 优化 RetrieveService 中的段落检索逻辑，按相似度分数排序结果

+++ src/main/java/com/tarzan/maxkb4j/module/dataset/vo/ParagraphVO.java
@@ -1,6 +1,5 @@
 package com.tarzan.maxkb4j.module.dataset.vo;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.tarzan.maxkb4j.module.dataset.entity.ParagraphEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -11,9 +10,9 @@ public class ParagraphVO extends ParagraphEntity {
     private String datasetName;
     private String documentName;
     private String hitHandlingMethod;
-    private Double similarity;
-    private Double directlyReturnSimilarity;
-    private Double comprehensiveScore;
+    private Float similarity;
+    private Float directlyReturnSimilarity;
+    private Float comprehensiveScore;
 
    public boolean isHitHandlingMethod(){
        return "directly_return".equals(hitHandlingMethod)&& similarity>=directlyReturnSimilarity;
